chore: moved the collaboration related api to instance_guest controller and wrote function for guest access token creation
