+++ lib/wraft_doc_web/controllers/instance_controller.ex
@@ -22,15 +22,9 @@ defmodule WraftDocWeb.Api.V1.InstanceController do
 
   require Logger
 
-  alias WraftDoc.Account
-  alias WraftDoc.Account.User
-  alias WraftDoc.AuthTokens
-  alias WraftDoc.AuthTokens.AuthToken
   alias WraftDoc.Client.Minio.DownloadError
   alias WraftDoc.Document
-  alias WraftDoc.Document.ContentCollaboration
   alias WraftDoc.Document.ContentType
-  alias WraftDoc.Document.CounterParties
   alias WraftDoc.Document.Instance
   alias WraftDoc.Document.Layout
   alias WraftDoc.Enterprise
@@ -1114,229 +1108,6 @@ defmodule WraftDocWeb.Api.V1.InstanceController do
     with %Instance{} = instance <- Document.show_instance(id, current_user),
          {:ok, _} <- Document.send_document_email(instance, params) do
       render(conn, "email.json", %{info: "Email sent successfully"})
-    end
-  end
-
-  @doc """
-   Share an instance.
-  """
-  swagger_path :invite do
-    post("/contents/{id}/invite")
-    summary("Share an instance")
-    description("Api to share an instance")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      content(:body, Schema.ref(:InviteDocumentRequest), "Share Request", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:Collaborator))
-    response(422, "Unprocessable Entity", Schema.ref(:Error))
-    response(401, "Unauthorized", Schema.ref(:Error))
-  end
-
-  @spec invite(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def invite(conn, %{"id" => document_id, "email" => _email} = params) do
-    current_user = conn.assigns.current_user
-
-    # send email in the end after the collaborator is created,
-    # didnt want to invoke sendgrid api all the time
-    # {:ok, %Oban.Job{}} <- Document.send_email(instance, invited_user, token)
-    with %Instance{state_id: state_id} = instance <-
-           Document.show_instance(document_id, current_user),
-         %User{} = invited_user <- Account.get_or_create_guest_user(params),
-         %ContentCollaboration{} = collaborator <-
-           Document.add_content_collaborator(current_user, instance, invited_user, params),
-         {:ok, %AuthToken{value: token}} <-
-           AuthTokens.create_document_invite_token(invited_user, state_id, params) do
-      # Just temporarily, ideally to be send via mailer
-      Logger.info("Invite token generated for user: #{token}")
-      render(conn, "collaborator.json", collaborator: collaborator, token: token)
-    end
-  end
-
-  @doc """
-  Verify document invite token.
-  """
-  swagger_path :verify_document_access do
-    get("/contents/{id}/verify_access/{token}")
-    summary("Verify document invite token")
-    description("Api to verify document invite token")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      token(:path, :string, "Invite token", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:VerifyDocumentInviteTokenResponse))
-    response(401, "Unauthorized", Schema.ref(:Error))
-    response(404, "Not found", Schema.ref(:Error))
-  end
-
-  @spec verify_document_access(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def verify_document_access(conn, %{"token" => token, "id" => document_id}) do
-    with {:ok, %{email: email, document_id: ^document_id, state_id: state_id}} <-
-           AuthTokens.check_token(token, :document_invite),
-         %User{} = user <- Account.get_user_by_email(email),
-         %ContentCollaboration{} = content_collaboration <-
-           Document.get_content_collaboration(document_id, user, state_id),
-         {:ok, %ContentCollaboration{}} <-
-           Document.accept_document_access(content_collaboration) do
-      render(conn, "check_token.json", token: token)
-    end
-  end
-
-  @doc """
-  Revoke document access.
-  """
-  swagger_path :revoke_document_access do
-    put("/contents/{id}/revoke_access/{collaborator_id}")
-    summary("Revoke document access")
-    description("Api to revoke document access")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      collaborator_id(:path, :string, "Collaborator id", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:Collaborator))
-    response(401, "Unauthorized", Schema.ref(:Error))
-    response(404, "Not found", Schema.ref(:Error))
-    response(422, "Unprocessable Entity", Schema.ref(:Error))
-  end
-
-  @spec revoke_document_access(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def revoke_document_access(conn, %{"id" => document_id, "collaborator_id" => collaborator_id}) do
-    current_user = conn.assigns.current_user
-
-    with %Instance{} = _instance <- Document.show_instance(document_id, current_user),
-         %ContentCollaboration{} = collaborator <-
-           Document.get_content_collaboration(collaborator_id),
-         %ContentCollaboration{} = collaborator <-
-           Document.revoke_document_access(current_user, collaborator) do
-      render(conn, "collaborator.json", collaborator: collaborator)
-    end
-  end
-
-  @doc """
-  List document instance collaborators.
-  """
-  swagger_path :collaborators do
-    get("/contents/{id}/collaborators")
-    summary("List document instance collaborators")
-    description("Api to list document instance collaborators")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:Collaborator))
-    response(401, "Unauthorized", Schema.ref(:Error))
-  end
-
-  @spec list_collaborators(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def list_collaborators(conn, %{"id" => document_id}) do
-    current_user = conn.assigns.current_user
-
-    with %Instance{} = instance <- Document.show_instance(document_id, current_user),
-         [%ContentCollaboration{} | _] = collaborators <- Document.list_collaborators(instance) do
-      render(conn, "collaborators.json", collaborators: collaborators)
-    end
-  end
-
-  @doc """
-  Update Collaborator role.
-  """
-  swagger_path :update_collaborator_role do
-    patch("/contents/{id}/collaborators/{collaborator_id}")
-    summary("Update Collaborator role")
-    description("Api to update collaborator role")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      collaborator_id(:path, :string, "Collaborator id", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:Content))
-    response(401, "Unauthorized", Schema.ref(:Error))
-  end
-
-  @spec update_collaborator_role(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def update_collaborator_role(
-        conn,
-        %{
-          "id" => document_id,
-          "collaborator_id" => collaborator_id
-        } = params
-      ) do
-    current_user = conn.assigns.current_user
-
-    with %Instance{} = _instance <- Document.show_instance(document_id, current_user),
-         %ContentCollaboration{} = collaborator <-
-           Document.get_content_collaboration(collaborator_id),
-         %ContentCollaboration{} = collaborator <-
-           Document.update_collaborator_role(collaborator, params) do
-      render(conn, "collaborator.json", collaborator: collaborator)
-    end
-  end
-
-  @doc """
-  Add counterpart to a contract document
-  """
-  swagger_path :add_counterparty do
-    post("/contents/{id}/add_counterparty")
-    summary("Add counterparty to a document")
-    description("Api to add counterpart to a document")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      counterparty(:body, Schema.ref(:CounterPartyRequest), "CounterParty", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:CounterPartyResponse))
-    response(401, "Unauthorized", Schema.ref(:Error))
-    response(404, "Not found", Schema.ref(:Error))
-    response(422, "Unprocessable Entity", Schema.ref(:Error))
-  end
-
-  @spec add_counterparty(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def add_counterparty(conn, %{"id" => document_id} = params) do
-    current_user = conn.assigns.current_user
-
-    with %Instance{} = instance <- Document.show_instance(document_id, current_user),
-         %CounterParties{} = counterparty <- Document.add_counterparty(instance, params) do
-      render(conn, "counterparty.json", counterparty: counterparty)
-    end
-  end
-
-  @doc """
-  Remove counterparty from a contract document
-  """
-  swagger_path :remove_counterparty do
-    PhoenixSwagger.Path.delete("/contents/{id}/remove_counterparty/{counterparty_id}")
-    summary("Remove counterparty from a document")
-    description("Api to remove counterparty from a document")
-
-    parameters do
-      id(:path, :string, "Instance id", required: true)
-      counterparty_id(:path, :string, "Counterparty id", required: true)
-    end
-
-    response(200, "Ok", Schema.ref(:CounterPartyResponse))
-    response(401, "Unauthorized", Schema.ref(:Error))
-    response(404, "Not found", Schema.ref(:Error))
-    response(422, "Unprocessable Entity", Schema.ref(:Error))
-  end
-
-  @spec remove_counterparty(Plug.Conn.t(), map) :: Plug.Conn.t()
-  def remove_counterparty(conn, %{"id" => document_id, "counterparty_id" => counterparty_id}) do
-    current_user = conn.assigns.current_user
-
-    with %Instance{} = _instance <- Document.show_instance(document_id, current_user),
-         %CounterParties{} = counterparty <-
-           Document.get_counterparty(document_id, counterparty_id),
-         %CounterParties{} = counterparty <- Document.remove_counterparty(counterparty) do
-      render(conn, "counterparty.json", counterparty: counterparty)
     end
   end
 end
